+++ docs/docs.go
@@ -1,0 +1,8 @@
+package docs
+
+import (
+	"embed"
+)
+
+//go:embed static
+var Docs embed.FS